+++ src/examples/cpp/protobuf_rpc/module/normal_rpc_co_server_module/service.cc
@@ -10,8 +10,8 @@ co::Task<aimrt::rpc::Status> ExampleServiceImpl::GetFooData(
     ::aimrt::protocols::example::GetFooDataRsp& rsp) {
   rsp.set_msg("echo " + req.msg());
 
-  AIMRT_INFO("Server handle new rpc call. req: {}, return rsp: {}",
-             aimrt::Pb2CompactJson(req), aimrt::Pb2CompactJson(rsp));
+  AIMRT_INFO("Server handle new rpc call. context:{}, req: {}, return rsp: {}",
+             ctx.ToString(), aimrt::Pb2CompactJson(req), aimrt::Pb2CompactJson(rsp));
 
   co_return aimrt::rpc::Status();
 }
@@ -22,8 +22,8 @@ co::Task<aimrt::rpc::Status> ExampleServiceImpl::GetBarData(
     ::aimrt::protocols::example::GetBarDataRsp& rsp) {
   rsp.set_msg("echo " + req.msg());
 
-  AIMRT_INFO("Server handle new rpc call. req: {}, return rsp: {}",
-             aimrt::Pb2CompactJson(req), aimrt::Pb2CompactJson(rsp));
+  AIMRT_INFO("Server handle new rpc call. context:{}, req: {}, return rsp: {}",
+             ctx.ToString(), aimrt::Pb2CompactJson(req), aimrt::Pb2CompactJson(rsp));
 
   co_return aimrt::rpc::Status();
 }