+++ src/interface/aimrt_module_cpp_interface/rpc/rpc_context.h
@@ -9,6 +9,7 @@
 #include "aimrt_module_c_interface/rpc/rpc_context_base.h"
 #include "aimrt_module_cpp_interface/util/string.h"
 #include "util/exception.h"
+#include "util/stl_tool.h"
 #include "util/string_util.h"
 #include "util/time_util.h"
 
@@ -67,6 +68,24 @@ class Context {
   }
   void SetSerializationType(std::string_view val) {
     SetMetaValue(AIMRT_RPC_CONTEXT_KEY_SERIALIZATION_TYPE, val);
+  }
+
+  std::string ToString() const {
+    std::stringstream ss;
+    ss << "timeout: " << timeout_ns_ / 1000000 << "ms, meta: {";
+    bool flag = true;
+    for (const auto& itr : meta_data_map_) {
+      if (flag)
+        flag = false;
+      else
+        ss << ",";
+
+      ss << "{" << itr.first << "," << itr.second << "}";
+    }
+
+    ss << "}";
+
+    return ss.str();
   }
 
  private:
@@ -187,6 +206,27 @@ class ContextRef {
   }
   void SetSerializationType(std::string_view val) {
     SetMetaValue(AIMRT_RPC_CONTEXT_KEY_SERIALIZATION_TYPE, val.data());
+  }
+
+  std::string ToString() const {
+    AIMRT_ASSERT(base_ptr_ && base_ptr_->ops, "Reference is null.");
+
+    std::stringstream ss;
+
+    auto timeout = base_ptr_->ops->get_timeout_ns(base_ptr_->impl);
+    ss << "timeout: " << timeout / 1000000 << "ms, meta: {";
+
+    aimrt_string_view_array_t keys = base_ptr_->ops->get_meta_keys(base_ptr_->impl);
+    for (size_t ii = 0; ii < keys.len; ++ii) {
+      if (ii != 0) ss << ",";
+      auto val = base_ptr_->ops->get_meta_val(base_ptr_->impl, keys.str_array[ii]);
+      ss << "{" << aimrt::util::ToStdStringView(keys.str_array[ii])
+         << "," << aimrt::util::ToStdStringView(val) << "}";
+    }
+
+    ss << "}";
+
+    return ss.str();
   }
 
  private: